chore(templates): anchor action bar at top of Prepare Invitation modal
Restructure the dialog shell as a flex column so the action bar stays
fixed in the header while only the tab content below scrolls. Moves
the checkbox + Cancel / Mark invited / Print / Send buttons out of
the bottom footer into the top-right of the header, next to the
speaker name + email line.

- Rename `PrepareInvitationFooter` → `PrepareInvitationActionBar` and
  tighten its layout for inline header use (smaller buttons, checkbox
  beside the buttons, `canSendReason` hint drops below).
- Dialog shell: `flex flex-col`; header is `shrink-0` with a bottom
  border, main area is `flex-1 overflow-auto`.
- Error message moves inline under the tab content (still visible
  when present, just no longer anchored at the bottom of the shell).
